Added the ADC target.h
I've used the betaflight firmware and checked the resources trough the CLI and it spit out this:
resource ADC_BATT 1 C04 
resource ADC_CURR 1 C05

This is needed because I think that my drone goes into failsafe/D_Shot_beacon mode because it detects 0v on the battery.

Would it be possible to add MFT-01 compatibility? It is a cheap optical/rangefinder module that should work up to 8m high, for GNSS denied enviroments.